Upgrades: Cleanup RabbitMQ / vhost
Commit b2624d4a2644edb5e8fe6060e8754d8fc781dc73 introduced per-service
vhosts in RabbitMQ as well as individual users. The OpenStack-related
exchanges and queues in the / vhost and the 'openstack' user are no
longer required and so are removed by this upgrade playbook
cleanup-rabbitmq-vhost.yml.

A release note is added mentioning the original change as well as the
upgrade tasks and the upgrade details are documented in the upgrade
guide.

A block of variables have also been added to the run-upgrade.sh script.
These are ncessary for the main function to execute, but were missed in
the prior commit. Without adding these (or modifying the referencing
code), the run-upgrade.sh script will not execute.

Change-Id: I9b1be43b887de5bcc41d83f19888aa0cd05f8133