Automate setting of glance_api_version
When integrating ceph w/ OpenStack, we suggest setting
glance_api_version to 2 in the cinder_backends section for the ceph
backend in question.  This allows cinder to create clones from glance
snapshots when glance uses RBD as a storage mechanism.

The problem is that glance_api_version needs to be set in [DEFAULT],
and not in the backend's section in cinder.conf.

This change adds a new user variable cinder_glance_api_version, and if
the environment has a cinder RBD backend and has glance configured to
use RBD it will then set glance_api_version in cinder.conf to 2.  If
these conditions are not met, then it will use version 1 (the default).

Change-Id: I242cd5b1e7eb47ee76c1f3cb0169e3b5d070959c
(cherry picked from commit 204c680f5202f1f93a1522c972a9757c58f012f6)