Add Kilo to Liberty upgrade scaffold
This commit provides scaffolding for future Kilo to Liberty work. While
incomplete, this commit should provide a space for the upgrade work to
be built open.

The `run_lock` function and its `RUN_TASKS` array are included as
the upgrade steps will be run through these methods, similar to the Juno
to Kilo process.

An upgrade guide section was added to the documentation so that further
work can be done there.

Since this commit is providing only a scaffold for the 'real' work to
follow, no release note has been added.

The work that needs to be done (in order of execution) is as follows:

    * Backup then update the environment/config files with Liberty
      values
    * Adjust the user secrets file for new secrets values.
    * Setup openstack and LXC hosts
    * Update container configurations
    * Restart container networking
    * Install/Upgrade infrastructure software (Galera, RabbitMQ)
    * Install Openstack Liberty code
    * Clean out RabbitMQ default vhost

Change-Id: Id1bb54039ca2fb15cbd3d199f777dbea063c9eec