Add upgrade config playbook and backup configs
This commit provides a playbook that initially will back up the
configuration directory, as well as documentation for that playbook. In
later commits, this playbook will be extended to change the existing
configuration in place.

The included release note is intended to serve for all configuration
related work, but not all such work is included yet in order to keep the
patch size manageable for reviewers.

Change-Id: I1dcc6639ec31262bbc9a8f3cfa6099a1c8e4d8dd