Update new user secrets in Liberty
The user-secrets-adjustments.yml playbook introduced for Kilo upgrades
is brought foward to Liberty in this commit. Since the task is logically
the same and new secrets variables were introduced, no changes from the
Kilo version are made. The documentation is updated to be more clear
about how existing secrets are handled.

Change-Id: I029b41d379d3cf3e073c7e004a9eb7986c56d1f1