Update user variables to Liberty names
In Liberty, some variables were renamed to better reflect upstream
decisions, such as Nova's default API versioning scheme. This change
provides a script to replace the old names with the new, while leaving
the rest of the user_variables.yml file intact.

A table of variable name changes for documentation can be generated with
the included make_rst_table.py script. This has to be run manually,
however, as this project doesn't use a Sphinx extension to include
arbitrary output in rst files, and the OpenStack publishing jobs use the
`setup.py build_sphinx` method instead of `make html`.

A test script for migrating the variables is provided as a check on the
migration functionality. However, this is not called as part of any
automated system as of this commit.

Change-Id: I0a898d2864d8f222af303f72f6119eb93d9e8488