Fix bugs when running run-upgrade.sh
Adjust the path for bootstrap-ansible.sh to use the correct variable
SCRIPTS_PATH.

Define the missing Ansible var root_dir in deploy-config-changes.yml and
rename it to repo_root_dir.

Modify add_new_neutron_env.py so that it updates the file
/etc/openstack_deploy/env.d/neutron.yml instead of the version being
carried in the repo. Pass into the script the path to the root of the
repo, this fixes the incorrect path calculated by the script.

Make fix_ceilometer_env.py executable so that it can be used by
run-upgrade.sh.

Change-Id: Ied10466c31c1e78d60efc17498daff0d24c2a380