Allow a list of files to be copied into the container cache
In order to better make use of OpenStack-CI's resources (apt mirrors, etc)
we need to be able to copy configuration files which have been implemented
on the host image into the containers.

This patch provides a facility to do this in a way that provides a new
facility for deployers to use, doesn't affect existing production
environments and is easy for us to consume in gate testing.

A new variable called 'lxc_container_cache_files' is implemented, which
is a list of dictionaries specifying the location of the file on the
deployment system, where to put it in the container cache, and what
attributes to assign the file.

An example of how to use it is included in the role's defaults as a comment.

Note:
This is a backport of https://review.openstack.org/289898 without the test
bits as the Liberty branch has no facility to use them.

Change-Id: Ia76b3a791239a07db1cb93636b7d1e23a6a6851c