Retry tempest cirros image creation
Creating the cirros image fails intermittently. To debug this I
increased the haproxy timeout and inserted some debug tasks into the
tempest_resources playbook. These revealed that the cause of the failure
is glance getting a 400 response from the cirros mirror when doing a
head request to ascertain the size of the image.

I cannot find fault with the OSA mechanisms and suspect the fault is an
intermittent one with the remote server. In which case we don't have
many options apart from to retry, which is what this patch does.

Cherry-Picked (ish) from:
openstack/openstack-ansible-os_tempest/dd12087021e5611b1be4a5a8a7512257e8844ba1

Change-Id: I9e9ac72acb6c8174db733405f6b468ac8ced4573