Fix command not found in scripts lib
The successorator originally incremented RETRY with ((RETRY++)).
Bashate decided that was an ambiguous redirect  (why?, I don't care
about the output, I just want the variable to incremented). As a result
we changed the increment statement to $((RETRY++)), this increments the
variable, then tries to execute it, so the log contains:

"scripts/scripts-library.sh: line 51: 1: command not found"

I was going to add _=$((RETRY++)) but its getting ugly, so I replaced
the while loop with a c-style for loop that allows the increment
operator in the third statement.

Change-Id: I3db03dcefbcbe2b4c51e414f9fd2fee73e16a62e
(cherry picked from commit 02c555ae0b4ee318c2e1b9637469cd99247f2976)