Add ansible fact clean up script to upgrade-utils
This change simply adds a fact cleanup script to the upgrade process
so that the system will gather fresh facts when upgrading. While this
will potentially add time to the upgrade process it will ensure we are
not running into issues from a previous release that may have been stored
in cached facts.

Change-Id: Ibae6e01d83ad4d53eafc63224c81e218de3ac013
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>