Gate: Allow jenkins user access to logs
The logs collected by the gate check are not all readable by the
jenkins user in OpenStack-CI. This patch ensures that it does.

Backport note:
The 'pywbem' cap is added to allow a gate pass. This will be
removed in a follow-on patch.

Change-Id: Ie8197c1aaccbbb052b7f58bb396fe1e1afb1b7cf
(cherry picked from commit a831ca82d1e2985a13478964781e1e42e21670e9)