Upgrade RabbitMQ on OSA upgrade
When upgrading OSA from an early Kilo release to 12.0.8, the RabbitMQ
version has likely changed and the 'Compare installed version of
RabbitMQ with new version variable' task will fail.

To avoid manual intervention in response to the failure, this change
introduces the RabbitMQ upgrade switch to both the run-upgrade.sh script and
the manual upgrade steps.

Change-Id: Id90cfb9e94cea1ff2598bc71211efa7ef4370445
Closes-Bug: #1563024